Add QA (Aqua/JET) groups to tableau sublibraries
Adds canonical QA test groups (Aqua.test_all + JET.test_package) to
OrdinaryDiffEqExplicitTableaus, OrdinaryDiffEqImplicitTableaus, and
OrdinaryDiffEqRosenbrockTableaus, following the existing sublibrary
pattern (ODEDIFFEQ_TEST_GROUP dispatch, test/qa environment with path
sources, [QA] on lts/1 in test_groups.toml). Part of SciML/.github#77.
