+++ lib/OrdinaryDiffEqExplicitTableaus/Project.toml
@@ -12,6 +12,7 @@ ODEProblemLibrary = "fdc4e326-1af4-4b90-96e7-779fcce2daa5"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 OrdinaryDiffEqExplicitRK = "9286f039-9fbf-40e8-bf65-aa933bdc4db0"
 OrdinaryDiffEqLowOrderRK = "1344f307-1e59-4825-a18e-ace9aa3fa4c6"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
@@ -27,8 +28,12 @@ DiffEqBase = "7"
 DiffEqDevTools = "3"
 ODEProblemLibrary = "1"
 OrdinaryDiffEq = "7"
+OrdinaryDiffEqExplicitRK = "2"
+OrdinaryDiffEqLowOrderRK = "2"
+Pkg = "1"
+Random = "<0.0.1, 1"
 Test = "<0.0.1, 1"
 julia = "1.10"
 
 [targets]
-test = ["DiffEqDevTools", "ODEProblemLibrary", "OrdinaryDiffEq", "OrdinaryDiffEqExplicitRK", "OrdinaryDiffEqLowOrderRK", "Random", "Test"]
+test = ["DiffEqDevTools", "ODEProblemLibrary", "OrdinaryDiffEq", "OrdinaryDiffEqExplicitRK", "OrdinaryDiffEqLowOrderRK", "Pkg", "Random", "Test"]
