+++ lib/OrdinaryDiffEqExplicitTableaus/test/qa/Project.toml
@@ -1,0 +1,20 @@
+[deps]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
+DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
+JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
+OrdinaryDiffEqExplicitTableaus = "3278f1b1-0f5c-4cde-98e0-ba5eb00db955"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[sources.DiffEqBase]
+path = "../../../DiffEqBase"
+
+[sources.OrdinaryDiffEqExplicitTableaus]
+path = "../.."
+
+[compat]
+Aqua = "0.8"
+DiffEqBase = "7"
+JET = "0.9, 0.10, 0.11"
+OrdinaryDiffEqExplicitTableaus = "2"
+Test = "1"
+julia = "1.10"
