+++ lib/OrdinaryDiffEqImplicitTableaus/Project.toml
@@ -8,6 +8,7 @@ DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
 
 [extras]
 DiffEqDevTools = "f3b72e0c-5b89-59e1-b016-84e28bfd966d"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [sources]
@@ -19,8 +20,9 @@ path = "../DiffEqDevTools"
 [compat]
 DiffEqBase = "7"
 DiffEqDevTools = "3"
+Pkg = "1"
 Test = "<0.0.1, 1"
 julia = "1.10"
 
 [targets]
-test = ["DiffEqDevTools", "Test"]
+test = ["DiffEqDevTools", "Pkg", "Test"]
