+++ lib/OrdinaryDiffEqRosenbrockTableaus/Project.toml
@@ -11,6 +11,7 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LinearSolve = "7ed4a6bd-45f5-4d41-b270-4a48e9bafcae"
 ODEProblemLibrary = "fdc4e326-1af4-4b90-96e7-779fcce2daa5"
 OrdinaryDiffEqRosenbrock = "43230ef6-c299-4910-a778-202eb28ce4ce"
+Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [compat]
@@ -21,11 +22,12 @@ LinearAlgebra = "1.10"
 LinearSolve = "3.46"
 ODEProblemLibrary = "1"
 OrdinaryDiffEqRosenbrock = "2"
+Pkg = "1"
 Test = "<0.0.1, 1"
 julia = "1.10"
 
 [targets]
-test = ["ADTypes", "DiffEqDevTools", "Enzyme", "LinearAlgebra", "LinearSolve", "ODEProblemLibrary", "Test", "OrdinaryDiffEqRosenbrock"]
+test = ["ADTypes", "DiffEqDevTools", "Enzyme", "LinearAlgebra", "LinearSolve", "ODEProblemLibrary", "Test", "OrdinaryDiffEqRosenbrock", "Pkg"]
 
 [sources.DiffEqDevTools]
 path = "../DiffEqDevTools"
