+++ lib/OrdinaryDiffEqRosenbrockTableaus/test/qa/Project.toml
@@ -1,0 +1,15 @@
+[deps]
+Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
+JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
+OrdinaryDiffEqRosenbrockTableaus = "b4bd8bb3-f80f-41d2-9b21-73a655b304b9"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[sources.OrdinaryDiffEqRosenbrockTableaus]
+path = "../.."
+
+[compat]
+Aqua = "0.8"
+JET = "0.9, 0.10, 0.11"
+OrdinaryDiffEqRosenbrockTableaus = "2"
+Test = "1"
+julia = "1.10"
